Fix: save, synchro and collection frequency of data fixed for the blue sensor + sensors names entry popup removed
I fixed the process of saving and persisting data into the database, and also the frequency of collecting data. For this part modifications are done in BluetoothLeService.kt.
User doesn't have to chose and type the name of his sensor anymore. for this part Modifications were done in SplashScreen.kt.